requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.6
pillow>=10.0.0
numpy>=1.24.0
pydantic>=2.0.0
aiofiles>=23.0.0
requests>=2.28.0